add modular css to home folder components
