+++ frontend/src/pages/Dashboard/Home/HomeContainer.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react'
 import Navbar from './Navbar'
 import { Outlet, useLocation, useNavigate } from "react-router-dom";
 import Header from './Header'
-
+import homeStyle from "./home.module.css"
 export default function HomeContainer() {
   const { pathname } = useLocation();
   const navigateTO = useNavigate();
@@ -12,11 +12,11 @@ export default function HomeContainer() {
     }
   }, [pathname, navigateTO]);
     return (
-        <div className='homeContainer'>
+        <div className={`${homeStyle.homeContainer}`}>
             {/* <h1>Instagram Dashboard</h1> */}
             <Header />
             <Navbar />
-            <div className="__outlet">
+            <div className={`${homeStyle.__outlet}`}>
                 <Outlet />
             </div>
         </div>
