+++ frontend/src/pages/Dashboard/setting/Privacy.js
